pyproject.toml
[project]
name = "campaign-generator"
version = "0.1.0"
description = "AI-driven creative automation for social ad campaigns"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "streamlit>=1.35.0",
    "pytest>=8.2.0",
    # Real adapter dependencies
    "openai>=2.2.0",
    "weaviate-client>=4.16.7",
    "boto3>=1.40.49",
    "Pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
include = ["app", "drivers", "tests"]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
